@brianrob 2020-03-18
Thanks @dsplaisted.  I appreciate the review.  @DamianEdwards, @KathleenDollard do you have comments on the performance characteristics of this change (size increase vs. throughput increase)?