@DamianEdwards 2020-03-20
My vote is to accept the full trade-off (i.e. pay the download/disk cost for the performance increase). We should of course keep monitoring the performance throughout the releases as we attack different things that impact these times such that we can adjust back if we feel we can make up the time gains through other means, but for now this gets my vote.