@brianrob 2020-03-20
Definitely agree that size is something we should continue to watch.  I'll look to see how easily we can automate this in the lab.

@KathleenDollard, any concerns on this one or should we merge?