@brianrob 2020-03-20
FYI, @billwert is currently working on a size on disk test for the SDK that will run alongside all of the other automated perf tests.