@KathleenDollard 2020-03-23
Sorry I missed this...Yup, let's take it.